Boost based on the n_pids field
This field is used as a proxy for popularity. When boosting on this
field and still using the manually constructed queries I encountered
problems with some searches, e.g. "dalsgaard svin", which ranked the
book "Danmark fauna" highest because it has lots of pids and an author
which almost matches svin: "Sven". Letting the edismax query parser
handle more of the query this particular problem went away.