Change boost values
Boosting on the individual fields in the qf part resulted in searches for an
author like Ma Jian becoming flooded with materials where the author or
title contained "ma". I've moved all the individual boost values into
the pf part and removed the creator part in bq and this results in much
more relevant results for Ma Jian (and Jian Ma since that's how his name
is catalogued most of the time - but that's another problem to solve).
The search for "dalsgaard svin" is now slightly worse in that the most
relevant result is now at second place but the overall results from our
search evaluation dataset are now better.